1. add udpbr & udputils 2. implement alive report

+++ recipes-apps/ledclient/ledclient/lcdclilib/lcdcli.c
@@ -18,25 +18,25 @@ int lcd_send_command(int x, int y, char* cmd)
 
 	if ((fd = socket(PF_UNIX, SOCK_STREAM, 0)) < 0) {
 		log_error("socket init fail!\n");
-		perror("socket");
 		ok = 0;
 	}
-	/*if( ok ){
+	//original marked
+	if( ok ){
 		memset(&addr, 0, sizeof(addr));
 		addr.sun_family = AF_UNIX;
 		strcpy(addr.sun_path, CLIENT_SOCK_FILE);
 		unlink(CLIENT_SOCK_FILE);
 		if (bind(fd, (struct sockaddr *)&addr, sizeof(addr)) < 0) {
-			perror("bind");
+			log_error("socket bind fail!\n");
 			ok = 0;
 		}
-	}*/	
+	}
 	if (ok) {
 		memset(&addr, 0, sizeof(addr));
 		addr.sun_family = AF_UNIX;
 		strcpy(addr.sun_path, SERVER_SOCK_FILE);
 		if (connect(fd, (struct sockaddr *)&addr, sizeof(addr)) == -1) {
-			perror("connect");
+			log_error("connect failed!\n");
 			ok = 0;
 		}
 	}
@@ -45,10 +45,10 @@ int lcd_send_command(int x, int y, char* cmd)
 		//strcpy (buff, "iccExchangeAPDU");
 		sprintf(buff, "%d:%d:%s", x, y, cmd);
 		if (send(fd, buff, strlen(buff)+1, 0) == -1) {
-			perror("send");
+			log_error("send error\n");
 			ok = 0;
 		}
-		log_debug("sent %s\n", buff);
+		//log_debug("sent %s\n", buff);
 	}
 	if (fd >= 0) {
 		close(fd);
